requirements.txt
watchdog==4.0.2
pystray==0.19.5
pillow==10.4.0
pywin32==306
